Move page model into separate file
